# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Terramaps API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tiles/{layer_id}/{z}/{x}/{y}.pbf": {
      "get": {
        "tags": [
          "MVT"
        ],
        "summary": "Get Tile",
        "description": "Get a vector tile for a specific layer at the given tile coordinates.\n\nFor order=0 (zip) layers: queries geography_zip_codes LEFT JOIN zip_assignments.\nFor order>=1 layers: queries pre-computed node geometries.\nData fields from data_field_config are included as flat numeric properties.",
        "operationId": "get_tile_tiles__layer_id___z___x___y__pbf_get",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Layer Id"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tiles/warm": {
      "get": {
        "tags": [
          "MVT"
        ],
        "summary": "Warm Cache",
        "operationId": "warm_cache_tiles_warm_get",
        "parameters": [
          {
            "name": "map_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}